# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WorkClub API
  version: 1.1.0
  description: |
    Multi-tenant API for agency client work, projects, tasks, time, proposals,
    invoices, teams, notifications and the restricted client portal.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: http://localhost:8888/api
    description: Local development
...
paths:
...
  /timeentry/start:
    post:
      tags: [Time]
      operationId: startTimer
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
  /timeentry/{id}/stop:
    patch:
      tags: [Time]
      operationId: stopTimer
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /timeentry/manual:
    post:
      tags: [Time]
      operationId: createManualTime
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ObjectId' }
...
  requestBodies:
    JsonObject:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
    Error:
      description: Request failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ObjectId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
...
    SuccessEnvelope:
      type: object
      required: [success, result]
      properties:
        success: { const: true }
        result: {}
    ErrorEnvelope:
      type: object
      required: [success, message]
      properties:
        success: { const: false }
        message: { type: string }
        details: {}
...
security:
  - bearerAuth: []